fix(deimv2): revert the AMP default to fp16 across train and eval
Follows the submodule change: DEIMv2 is an fp16 recipe and we had no evidence
for forcing bf16 on it. The predictor's eval dtype tracks the same
KCD_AMP_DTYPE knob so a run trained under an explicit setting is scored the
same way -- gen003's export/score script pins bfloat16 for exactly that reason,
since gen003 is the one checkpoint that trained in bf16.

Outputs are still cast to float32 in _forward regardless of dtype. That fix is
independent of this one and stays: it is required for any non-fp32 autocast,
and it is what job 490 died without.

+++ kwcoco_detector_kit/trainers/deimv2.py
@@ -823,16 +823,17 @@ class DEIMv2Predictor:
         # cores (matters a lot for tiled eval's ~55 forward passes/image).
         # Disable with KCD_EVAL_AMP=0. CPU path stays fp32.
         #
-        # The dtype MUST match what training used, and training now defaults
-        # to bfloat16 (see tpl/DEIMv2/engine/solver/det_engine.py). Evaluating
-        # bf16-trained weights under fp16 reintroduces exactly the ~65504
-        # ceiling that bf16 was adopted to escape, on a stack with a known
-        # history of activation excursions. KCD_AMP_DTYPE overrides both.
+        # The dtype tracks whatever training used, which defaults to fp16 --
+        # DEIMv2's own recipe (see tpl/DEIMv2/engine/solver/det_engine.py).
+        # KCD_AMP_DTYPE overrides both together, so a run trained under an
+        # explicit bf16 setting is scored the same way by passing the same
+        # value. Outputs are cast back to float32 in _forward regardless,
+        # since numpy has no bfloat16.
         self._use_amp = (
             str(device).startswith("cuda")
             and os.environ.get("KCD_EVAL_AMP", "1") != "0"
         )
-        self._amp_dtype_name = os.environ.get("KCD_AMP_DTYPE", "bfloat16").lower()
+        self._amp_dtype_name = os.environ.get("KCD_AMP_DTYPE", "float16").lower()
 
     @profile
     def _forward(self, im, sz):
